+++ tools/pylint-extensions/azure-pylint-guidelines-checker/tests/test_files/missing_dep_poetry_test_pkg/acceptable.py
@@ -1,0 +1,10 @@
+# ------------------------------------
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+# ------------------------------------
+
+# Acceptable: typing_extensions is in pyproject.toml [tool.poetry.dependencies]
+import typing_extensions
+
+# Acceptable: requests is in pyproject.toml [tool.poetry.dependencies]
+import requests